Add an error parameter for translation unit creation
Although libclang does not currently expose the reason for failure
during translation unit creation (file I/O error, AST file created from
a different clang version, etc) it would be nice if it did so in the
future. Provide an API that allows this information to be exposed in the
future without breaking source compatibility.