Have cursors and types hold a reference to the translation unit
This keeps the CXTranslationUnit (and by extension the CXIndex) valid as
long as a reference is held to an object that depends on it. To avoid a
retain cycle the translation unit does not retain its PLClangCursor.

This change frees the client from memory management concerns for CX
objects in most cases. Reparsing a translation unit and some code
completion operations can still invalidate existing objects associated
with a translation unit, for now this is something clients will need to
be aware of.