Lazily load all properties for a cursor
Profiling revealed the PLClangCursor initializer's creation of the USR,
spelling, and display name strings to be a significant source of CPU use
when visiting the cursors of a translation unit. In situations where
only a subset of the cursors are of interest, such as those not in
system headers, the work done to create these strings in the initializer
slows down the application unecessarily. Lazily load these properties as
well to remove this overhead.