Defer reading the expansion location until needed
When visiting all the cursors in a translation unit it is common to skip
over those in system headers. Profiling revealed
clang_getExpansionLocation() to be a relatively expensive operation that
is not needed to query clang_Location_isInSystemHeader(). Defering
reading of the expansion location until an associated property is read
allows isInSystemHeader to be read without incurring the cost of
obtaining the expansion location for the tens of thousands of
declarations in system headers.