Use the project's deployment target when building Clang
This is accomplished by preserving the MACOSX_DEPLOYMENT_TARGET
environment variable in the cleaned environment. In the absence of an
explicit -mmacosx-version-min argument Clang will read the value of this
environment variable determine the deployment target.

The --enable-libcpp option is also passed to LLVM's configure script.
With a deployment target of 10.8 the default C++ standard library is
libstdc++, and the 4.6 version of libstdc++ that is included in Xcode is
missing features required to build Clang.